refactoring listagem
