+++ WebContent/assets/veiculo.css
@@ -1,0 +1,11 @@
+body {
+	padding: 0 2em;
+}
+
+table {
+	width: 100%;
+}
+
+.pure-menu-horizontal {
+	background: aliceblue;
+}
