+++ src/br/com/fiap/ltp/nac02/questao1/veiculo/Veiculo.java
@@ -1,23 +1,30 @@
 package br.com.fiap.ltp.nac02.questao1.veiculo;
 
 public class Veiculo {
-	
+	private String id;
 	private String modelo;
 	private String placa;
 	private String ano;
 	private String motor;
 	private String placaNova;
-	
-	
+
 	public Veiculo() {
-		
+
 	}
 
-	public String getPlacaNova(){
+	public String getId() {
+		return id;
+	}
+
+	public void setId(String id) {
+		this.id = id;
+	}
+
+	public String getPlacaNova() {
 		return placaNova;
 	}
 
-	public void setPlacaNova(String placaNova){
+	public void setPlacaNova(String placaNova) {
 		this.placaNova = placaNova;
 	}
 
@@ -25,40 +32,32 @@ public class Veiculo {
 		return modelo;
 	}
 
-
 	public void setModelo(String modelo) {
 		this.modelo = modelo;
 	}
-
 
 	public String getPlaca() {
 		return placa;
 	}
 
-
 	public void setPlaca(String placa) {
 		this.placa = placa;
 	}
-
 
 	public String getAno() {
 		return ano;
 	}
 
-
 	public void setAno(String ano) {
 		this.ano = ano;
 	}
-
 
 	public String getMotor() {
 		return motor;
 	}
 
-
 	public void setMotor(String motor) {
 		this.motor = motor;
 	}
-	
 
 }
